a bit of restructuring of uasset reading in prep for a way to handle message assets + hot fix for gift items when item pool is not combined that makes sure all items are available somewhere
