+++ base_classes/message_logic.py
@@ -1,0 +1,7 @@
+from base_classes.uasset import UAsset
+from util.binary_table import Table
+from base_classes.script_logic import readBinaryTable, writeBinaryTable, writeFolder
+
+class Message_Uasset(UAsset):
+    def __init__(self, binaryTable: Table):
+        UAsset.__init__(self, binaryTable)
